Renaming src/page -> src/pageComponents
