+++ template/app/signature-mint/index.tsx
@@ -3,13 +3,13 @@
 import dynamic from 'next/dynamic';
 import Banner from '../../src/components/banner/banner';
 import Header from '../../src/components/header/Header';
-import Guide from '../../src/pages/signature-mint/Guide';
+import Guide from '../../src/pageComponents/signature-mint/Guide';
 
 // Because the mint page relies so heavily on client-side state, without disabling SSR
 // for its internals we get annoying hydration errors. A future enhancement would be to
 // read token metadata through a provider that is available server-side.
 const SignatureMint = dynamic(
-  async () => import('../../src/pages/signature-mint/ContractDemo').then((mod) => mod),
+  async () => import('../../src/pageComponents/signature-mint/ContractDemo').then((mod) => mod),
   {
     ssr: false,
   },
