--- template/pages/README.md
@@ -1,1 +1,0 @@
-# Leave this file here.. otherwise Next will try to compile ./src/pages and error
