+++ template/src/pageComponents/buy-me-coffee/ContractDemo.tsx
