+++ template/src/pageComponents/home/Guide.tsx
