+++ template/src/pageComponents/mint/ContractDemo.tsx
