+++ template/src/pageComponents/mint/Guide.tsx
