+++ template/src/pageComponents/mint/NotConnected.test.tsx
