+++ template/src/pageComponents/mint/NotConnected.tsx
