+++ template/src/pageComponents/mint/SwitchNetwork.tsx
