+++ template/src/pageComponents/signature-mint/Guide.tsx
